# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Batchmates API",
        "description": "Multi-tenant alumni donation and campaign management system API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/campaigns/{campaign}/approve": {
            "post": {
                "tags": [
                    "Campaign Approval"
                ],
                "summary": "Approve campaign",
                "description": "Committee member approves campaign",
                "operationId": "19be330eccd3ded1517b0cc27391ce93",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Campaign looks good, approved."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign approved"
                    },
                    "403": {
                        "description": "Not authorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{campaign}/reject": {
            "post": {
                "tags": [
                    "Campaign Approval"
                ],
                "summary": "Reject campaign",
                "description": "Committee member rejects campaign with reason",
                "operationId": "cf4c5fd4199f96a2e2fa35ac4c7a68a4",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Missing required documentation"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Please provide budget breakdown"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign rejected"
                    },
                    "403": {
                        "description": "Not authorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{campaign}/complete": {
            "post": {
                "tags": [
                    "Campaign Approval"
                ],
                "summary": "Mark campaign as completed",
                "description": "Mark campaign as completed when goal is reached",
                "operationId": "ded5b30afe7f0ca106e33540f37913c5",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign marked as completed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{campaign}/close": {
            "post": {
                "tags": [
                    "Campaign Approval"
                ],
                "summary": "Close campaign early",
                "description": "Manually close campaign before reaching goal",
                "operationId": "0ecb3fd63debef5841634c8096b94f5f",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Project requirements changed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign closed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaignStats/pending-review": {
            "get": {
                "tags": [
                    "Campaign Approval"
                ],
                "summary": "Get campaigns pending review",
                "description": "List all campaigns waiting for committee approval",
                "operationId": "4e69bec24abe3ccab201e763179c8245",
                "responses": {
                    "200": {
                        "description": "List of pending campaigns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{campaign}/history": {
            "get": {
                "tags": [
                    "Campaign Approval"
                ],
                "summary": "Get campaign approval history",
                "description": "View all approval actions for a campaign",
                "operationId": "883f96639660469e788d2497552e5a7e",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approval history"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "Use this for MOBILE testing (Token-based)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "CookieCsrfToken": {
                "type": "apiKey",
                "description": "Use this for WEB testing (Session-based). Value must be the XSRF-TOKEN cookie value.",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}